Added second page
I added a template for the second page, all I need to do to make the tutorial pages is add in the text, and I should be done the page.
I also started on trying to fix the feedback I got, but did not have enough time to do that.