Added first tutorial
I finished the first tutorial and changed up some of the fonts for the code to make it look good.
I added a copy to clipboard button. This took me a long time and I should be able to copy this over easily and finish the next tutorials in much less time